Fix arithmetic overflow warnings using sqrtf instead of sqrt

+++ 3DPerspective/Main.cpp
@@ -11,12 +11,12 @@ void line(Screen& screen, float x1, float y1, float x2, float y2)
 	float dx = x2 - x1;
 	float dy = y2 - y1;
 
-	float length = sqrt(dx * dx + dy * dy); // apply Pythagorean theory
+	float length = sqrtf(dx * dx + dy * dy); // apply Pythagorean theory
 	float angle = atan2(dy, dx);
 
 	for (float i = 0; i < length; i++)
 	{
-
+		
 	}
 }
